relation library => books

+++ src/feature/library/library.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { EntityRepository, EntityManager } from '@mikro-orm/core';
+import { EntityRepository, EntityManager, FindOptions } from '@mikro-orm/core';
 import { Library } from './entity/library.entity';
 import { UpdateLibraryDto } from './dto/update-library.dto';
 import { CreateLibraryDto } from './dto/create-library.dto';
 import { NotFoundLibraryExceptionDto } from './dto/not-found-library-exception.dto';
+import { Book } from '../books/entity/books.entity';
 
 @Injectable()
 export class LibraryService {
@@ -19,14 +20,10 @@ export class LibraryService {
   }
 
   async findOne(id: number): Promise<Library | null> {
-    const library = await this.libraryRepository.findOne({ id });
+    const library = await this.libraryRepository.findOne({ id }, { populate: ['books'] });
     if (!library) throw new NotFoundLibraryExceptionDto();
     return library;
   }
-
-  async findBooks(libraryId: number) {
-    return (await this.findOne(libraryId)).id;
-   }
 
   async create(libraryData: CreateLibraryDto): Promise<Library> {
     const library = this.libraryRepository.create(libraryData);
@@ -39,16 +36,17 @@ export class LibraryService {
     libraryData: UpdateLibraryDto,
   ): Promise<Library | null> {
     const library = await this.findOne(id);
+    if (!library) throw new NotFoundLibraryExceptionDto();
 
-    const newLibrary: Library = libraryData as Library;
-
-    this.libraryRepository.assign(library, newLibrary);
+    this.libraryRepository.assign(library, libraryData);
     await this.em.flush();
     return library;
   }
 
   async delete(id: number): Promise<boolean> {
     const library = await this.findOne(id);
+    if (!library) throw new NotFoundLibraryExceptionDto();
+
     await this.em.removeAndFlush(library);
     return true;
   }
